Déplacement des scripts dans leur dossier propre
